Brian Ko Summary + Potential Findings Section
Created a new file named bko_summary.ipynb in the visualization folder. Started to create a general summary table but I realized that we already have a decent one in the main folder so I created a table focusing on the average glucose, bmi, and blood pressure for either Type 1 or 2 diabetes. I also realized that we may be considering using the pregnancies variable, so I changed the number of pregnancies to a factor and found the number of data available for each number of pregnancies. Then I found the average glucose, bmi, and blood pressure for each number of pregnancies. Finally, I created 7 potential statements for the "Such findings will be able to create a much bigger impact in the future such as..." We can later discuss as a group which one of these may be the best for our project.